requirements.txt
streamlit
pillow
opencv-python
pandas
imageio
ultralytics

torch
torch-vision